+++ c/src/lib/libcpu/powerpc/mpc860/console-generic/console-generic.c
@@ -102,8 +102,8 @@ m860_get_brg_cd (int baud)
 /*  at any time since the OS started. It needs to be fixed. FIXME */
 unsigned char m860_get_brg_clk(int baud)
 {
-  static short brg_spd[4];
-  static char  brg_used[4];
+  static int  brg_spd[4];
+  static char brg_used[4];
   int i;
 
   /* first try to find a BRG that is already at the right speed */
